backend/requirements.txt
scapy  # For data collection
pandas  # For data preprocessing
transformers  # For LLaMa model (Hugging Face)
web3  # For blockchain integration
flask

google-cloud-storage
transformers
datasets